requirements.txt
beautifulsoup4>=4.6.0
bs4>=0.0.1
certifi>=2018.1.18
cfscrape>=1.9.4
chardet>=3.0.4
idna>=2.6
PyExecJS>=1.5.1
python-dateutil>=2.6.1
requests>=2.18.4
six>=1.11.0
urllib3>=1.22
